requirements.txt
streamlit
joblib
numpy
Pillow
matplotlib
scikit-learn
